More refactoring in perparation for issue #14. Merging ParagraphRenderer class into ParagraphElement

+++ pdf/src/main/groovy/com/craigburke/document/builder/render/RowElement.groovy
@@ -19,7 +19,7 @@ class RowElement implements Renderable {
     Row node
     List<CellElement> cellElements = []
 
-    RowElement(Row row, int startX) {
+    RowElement(Row row, float startX) {
         this.node = row
         this.startX = startX
